feat: 将标签分类工具组件化
