+++ src/components/book-list/index.vue
@@ -5,21 +5,7 @@
     <a-row :gutter="20" align="stretch" style="overflow-x: hidden">
       <!-- 标签工具栏 -->
       <a-col :span="24" v-if="renderData.length > 0">
-        <a-space direction="horizontal" wrap :v-if="tagsData.length > 0">
-          <span :style="{ color: 'var(--color-neutral-8)' }">标签：</span>
-          <template v-if="tagId ?? 0 > 0">
-            <a-tag :color="curTag?.Color" :style="{ cursor: 'pointer' }">
-              {{ curTag?.Text }}({{ curTag?.Count }})
-            </a-tag>
-            <a-tag @click="CheckTag(undefined)"><template #icon><icon-close /></template> </a-tag>
-          </template>
-          <template v-else>
-            <a-tag v-for="t of tagsData" :color="t.Color" :key="t.id" :style="{ cursor: 'pointer' }"
-              @click="CheckTag(t.id)">
-              {{ t.Text }}({{ t.Count }})
-            </a-tag>
-          </template>
-        </a-space>
+        <TagList :tagid="props.tagid" :Api="props.Api" @change="renderData = $event" />
       </a-col>
       <a-divider />
       <a-col :span="24">
@@ -41,7 +27,8 @@
                     </a-doption>
                     <a-doption>
                       <template #icon> <icon-pen /> </template>
-                      <template #default><a-button type="text" @click="curEditBookId=item.BookId">修改元数据</a-button></template>
+                      <template #default><a-button type="text"
+                          @click="curEditBookId = item.BookId">修改元数据</a-button></template>
                     </a-doption>
                   </template>
                 </a-dropdown>
@@ -53,18 +40,17 @@
     </a-row>
   </a-spin>
 
-  <EditBookInfo :visible="curEditBookId != 0" :bookId="curEditBookId" @cancel="curEditBookId = 0"/>
+  <EditBookInfo :visible="curEditBookId != 0" :bookId="curEditBookId" @cancel="curEditBookId = 0" />
 </template>
 
 <script lang="ts" setup>
 import { PropType, ref } from 'vue';
 import { useRouter } from 'vue-router';
 import { ListQueryApi, deleteABook } from '@/api/library';
-import { getTagHasBook, Tag } from '@/api/tag';
 import { Book } from '@/types/book';
 import useRequest from '../../hooks/request';
 import BookCover from '../book-cover/index.vue';
-
+import TagList from '../tag-toolbar/list.vue';
 import EditBookInfo from '../book-info/edit.vue';
 
 // 入参
@@ -77,8 +63,7 @@ const props = defineProps({
     required: true
   }
 });
-const tagId = ref(props.tagid);
-const curTag = ref<Tag>();
+
 const defaultValue: Book[] = new Array().fill({});
 const curEditBookId = ref(0);
 
@@ -90,17 +75,6 @@ const { loading, response: renderData } = useRequest<Book[]>(
   defaultValue
 );
 
-//标签相关逻辑
-const { response: tagsData } = useRequest<Tag[]>(getTagHasBook);
-function CheckTag(id: number | undefined) {
-  tagId.value = id;
-  loading.value = true;
-  props.Api(id).then(result => {
-    loading.value = false;
-    renderData.value = result.data as any;
-    curTag.value = tagsData.value.filter(t => t.id === tagId.value)[0];
-  })
-}
 
 const router = useRouter();
 const goto = (bookid: number) => {
@@ -115,7 +89,7 @@ function DeleteABook(bookid: number) {
     for (let i = 0; i < renderData.value.length; i++) {
       if (renderData.value[i].BookId == bookid) {
         loading.value = false;
-        renderData.value.splice(i,1);
+        renderData.value.splice(i, 1);
         return;
       }
     }
